+++ core/data/src/main/java/com/wap/wapp/core/data/repository/event/EventRepository.kt
@@ -1,7 +1,8 @@
 package com.wap.wapp.core.data.repository.event
 
 import com.wap.wapp.core.model.event.Event
+import java.time.LocalDate
 
 interface EventRepository {
-    suspend fun getNowMonthEvents(): Result<List<Event>>
+    suspend fun getMonthEvents(date: LocalDate): Result<List<Event>>
 }
