+++ core/domain/src/main/java/com/wap/wapp/core/domain/usecase/event/GetEventsUseCase.kt
@@ -2,10 +2,12 @@ package com.wap.wapp.core.domain.usecase.event
 
 import com.wap.wapp.core.data.repository.event.EventRepository
 import com.wap.wapp.core.model.event.Event
+import java.time.LocalDate
 import javax.inject.Inject
 
 class GetEventsUseCase @Inject constructor(
     private val eventRepository: EventRepository,
 ) {
-    suspend operator fun invoke(): Result<List<Event>> = eventRepository.getNowMonthEvents()
+    suspend operator fun invoke(date: LocalDate): Result<List<Event>> =
+        eventRepository.getMonthEvents(date)
 }
