+++ core/network/src/main/java/com/wap/wapp/core/network/source/event/EventDataSource.kt
@@ -1,7 +1,8 @@
 package com.wap.wapp.core.network.source.event
 
 import com.wap.wapp.core.network.model.event.EventResponse
+import java.time.LocalDate
 
 interface EventDataSource {
-    suspend fun getNowMonthEvents(): Result<List<EventResponse>>
+    suspend fun getMonthEvents(date: LocalDate): Result<List<EventResponse>>
 }
