+++ feature/notice/src/main/java/com/wap/wapp/feature/notice/BottomSheetContent.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.wap.designsystem.WappTheme
 import com.wap.designsystem.component.Loader
-import com.wap.wapp.core.domain.util.DateUtil.Companion.MONTH_DATE_START_INDEX
+import com.wap.wapp.feature.notice.DateUtil.Companion.MONTH_DATE_START_INDEX
 import com.wap.wapp.core.model.event.Event
 import java.time.format.DateTimeFormatter
