+++ feature/notice/src/main/java/com/wap/wapp/feature/notice/NoticeViewModel.kt
@@ -12,11 +12,13 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
+import java.time.LocalDate
 import javax.inject.Inject
 
 @HiltViewModel
 class NoticeViewModel @Inject constructor(
     private val getEventsUseCase: GetEventsUseCase,
+    private val dateUtil: DateUtil,
 ) : ViewModel() {
 
     private val _signUpEventFlow = MutableSharedFlow<NoticeEvent>()
@@ -25,16 +27,19 @@ class NoticeViewModel @Inject constructor(
     private val _events = MutableStateFlow<EventsState>(EventsState.Loading)
     val events: StateFlow<EventsState> = _events.asStateFlow()
 
+    private val _selectedDates = MutableStateFlow<LocalDate>(dateUtil.generateNowDate())
+    val selectedDate: StateFlow<LocalDate> = _selectedDates.asStateFlow()
+
     init {
-        getNowMonthEvents()
+        getMonthEvents()
     }
 
     fun event(event: NoticeEvent) = viewModelScope.launch { _signUpEventFlow.emit(event) }
 
-    fun getNowMonthEvents() {
+    fun getMonthEvents() {
         _events.value = EventsState.Loading
         viewModelScope.launch {
-            getEventsUseCase().fold(
+            getEventsUseCase(_selectedDates.value).fold(
                 onSuccess = { _events.value = EventsState.Success(it) },
                 onFailure = { _events.value = EventsState.Failure(it) },
             )
