+++ feature/notice/src/test/java/com/wap/wapp/feature/notice/DateUtilTest.kt
@@ -1,6 +1,5 @@
 package com.wap.wapp.feature.notice
 
-import com.wap.wapp.core.domain.util.DateUtil
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.Test
 import java.time.LocalDate
